multi-agents-cdk/requirements.txt
aws-cdk-lib==2.176.0
constructs>=10.0.0,<11.0.0
aws-cdk-lib>=2.150.0
constructs>=10.3.0
cdk-nag>=2.28.158
boto3>=1.34.150
python-dotenv>=1.0.1